Add PacketConnectionSplicer to duplicate packets into multiple destinations
This class allows to duplicate packets into multiple queues and takes care over
incrementing the reference counter (see ManagedInstance). It was a desing
decision to only allow adding new destinations and not to remove existing
destinations, because otherwise one has to use a mutex in the hot path,
which would harm performance.

git-svn-id: file:///Users/braun/svn/vermont/branches/vermont-dynamic-config@1025 aef3b71b-58ee-0310-9ba9-8811b9f0742f