Implement method send which will wait until the Source is connected
We use a CountingSemaphore here because we don't want to use a Mutex/Semaphore
in the hotpath for send. We only check if we are connected and use the
destination we gathered inside this check (this is all lockfree).

If someone disconnects the source in the meantime, the Packet would be send
to the old destination instead of the new one.

But at first, the Source has to sleep until it is connected to a destination.
This has to be race free, so we could asure that dest is not NULL. For this,
a CountingSemaphore initialized with 1 is used.
In the method sleepUntilConnected we try to decrement the semaphore by 2,
which will block if not connected. In connect() we do sem +1 and in
disconnect() sem -1. This prevents a race between sleepUntilConnected()
and disconnect(), because disconnect() will always succeed even if we are
waiting in sleepUntilConnected().


git-svn-id: file:///Users/braun/svn/vermont/branches/vermont-dynamic-config@1063 aef3b71b-58ee-0310-9ba9-8811b9f0742f