Ported Oracle DB Reader to new generic reader
The Oracle DB Reader now uses the generic reader classes and
can now read the new table format (similar than the MYSQL reader).
The code is not tested due to missing oracle libraries on the
development machine. Next commit should fix problems (if there are
any), and should probably be merged with this one ...)