Removed static IP_HEADER_OFFSET cmake option
Previous versions of VERMONT required a proper compile time option
that tells VERMONT about the size of the layer 2 headers.
This means that VERMONT needed to be recompiled if it was
compiled with the default option and needed to be run on ethernet
traffic with VLAN tags.
The current commit enables dynamic detection of the pcap link type
(Ethernet, BSD loopback, ...) and adopts the IP header offset
accordingly.
Furthermore, the code checks for ethernet VLAN tags for each
packet. This allows to analyze traffic that mixes VLAN-tagged
and untagged traffic.