Do not copy Layer 2 headers
PacketAggregator requires a static netHeaderOffset for its
highly optimized aggregation process. When we enable dynamic
link layer detection, we cannot know the offset of the IP header
before we opened the capturing device (or see the layer 2 header
in case of VLANs).
In order to maintain a fixed offset into the packet data structure
we do not copy the unknown sized layer 2 header, but only start
to copy data at the beginning of the IP header. Hence, the offset
is now static (it is zero for all packets)