Readded layer2 header information
The previous commits removed the layer2 header information from the
packets. However, this information is used by some modules, such as
the PCAPExporterBase. This commit restores the layer2 header, which
is now still saved in the Packet structure and can be accessed by a
separate pointer. Modules have been rewritten to use this pointer
if they need to.