Remove default IpfixSender record limitation
Previous versions of vermont had a default record limit rate
of 5000 records per second that could be sent at maximum. Vermont
would start sleeping when this limit is reached, leading to filled
up buffers and therefore packet loss.

In networks with high numbers of flows, this default parameter could
lead to undesired effects that are not entirely obvious. This commit
changes the default value to no limitations and removes calls to
gettimeofday when no limit is applied.

fixes #15