IpfixFileWriter: check for committed bytes
ipfixlolib did not check if any bytes have been committed to
the sendbuffer before writing a new message to the IPFIX files.
Instead it prepares an IPFIX header with length 16 and a
message which does not contain any data sets (which violates
the definition of a IPFIX message which says:

"
An IPFIX Message consists of a Message Header, followed by one or
more Sets.
"

fixes: #18