Add ZMQ IpfixReceiver subclass
Add support to receive Ipfix/Netflow payloads over a ZMQ connection.
The support is optional and disabled by default. If enabled, adds
a build dependency on libczmq. CZMQ is a high-level API that builds
on libzmq, and exposes easy and nice to use APIs to communicate using
the ZeroMQ protocol.
For more information about ZeroMQ, please see:
http://zeromq.org/
https://github.com/zeromq/czmq