Rename and simplify ipfix_prepend_header function
Is now called ipfix_update_header to reflect its function.
Does not need to be passed the data_length as in all cases
it just uses the committed length.

It could be further improved to calculate the length directly from the
vectors as the first vector should contain the header, so need for the
special case.