NFV9 support in ipfixlolib
NFV9 and IPFIX are very similar, the later really just evolution and
standardised version of the former. Only major difference is the header
fields and Template IDs. (Options also differ but library does not
support them for either protocol yet).

Added the relative timestamp to the header. (Future task to add option
to the Vermont exporter module to replace/add relative timestamps that
most V9 expect to receive instead of/as well as the absolute times)

Count field was easy as it's just the records (of types) in a message,
and library sends data an template records in separate messages.

Sequence count was harder, as it's the number of messages per collector
connection session, and as template manage is complex compared to data
messages, as sessions come and and are managed this state spread through
the code.

ipfix_send_template() is crying out to be refactored and simply the
tracking of internal state across it. As I become more comfortable with
code I'll most likely attempt such a refactor.