Support 16 bits tcpControlFlags in the Ipfix Aggregator
The new Ipfix RFCs (7011 and 7012, see
http://www.iana.org/assignments/ipfix/ipfix.xhtml) have changed
the size of the tcpControlFlags from 1 byte to 2 bytes.
Some part of the code, like the ipfixlolib, already support it,
but the Ipfix Aggregator asserts when the field is not 1 byte long.
The RFC and IANA page specifically say that in case an exporter does
not support the upper flags, then 1 byte should be exported. So, in
Vermont, we have to support both cases in ingress. But internally, in
the aggregator, store everything in 2 bytes format. So if there is no
aggregator and an exporter sends 1 byte field, we export 1 byte. But
if there is an aggregator, convert to 2 bytes.

Signed-off-by: Luca Boccassi <lboccass@brocade.com>

Conflicts:
	src/modules/ipfix/IpfixCsExporter.hpp