Add Cmake option to disable UDP connect
Calling connect on a UDP socket will fail if there are no interfaces
available, and the exporter module will be left in an errored state and
will not recover. Add an option to disable the connect call at build
time, and also change from using writev to sendmsg for UDP sockets
(functionality is equivalent, and resulsting system calls are similar
too).