On send, use right call to check if dest exists
When a module tries to send a packet to the next in the chain, it checks
to see whether there is a destination or the module is the last of the
chain. It was using an half-unused boolean, which was set to false if
the module at startup had no successor and never updated if on a reload
a successor is added. Use the proper check and drop this variable.