Remove ConfigTester from VermontTest
It seems horrible broken, with assumptions about locations of config
files and output files that don't work with out of source builds,
requirements for PCAP sizes that can't be met, and unsupported
connection based filters that all result in crashes.
It only seems to test connection filter at any rate so not much of a
lose.