Better ordering of msg_ types.
Normal behaviour such as vermont confirming it was up and running were
being reported with ERR severity level. This makes them warnings and
errors actually error.

It's still not ideal, with MSG_VDEBUG not mapping well to syslog levels
and forcing all the other levels down one. Ideally the whole code base
would be switched to just using the syslog levels, but that would
be 100/1000s of lines of (admittedly easily done with sed) diff.