Allow active/inactive timer to be set to 0
Do not rais an exception if mix/maxBufferTime are set to 0.
Flow cache has a "now" class variable, but was missing initialisation.
When checking if a flow is expired, check if smaller or equal to the
limit, for the corner case where the values are both 0.
If the flow has no flowStart/End field, initialise the internal
variable used to expire the cache to the current Unix time.