Compare active/inactive timer with wall clock
Currently the active/inactive timers are compare with the flowStartTime
of the last report that was added to a cache (bucket). This means that
the last report will be forever in the cache if the last report
flowStartTime was smaller than the active timeout.
Use instead the wall clock to decide if a flow is expired or not.