ipfixlolib DTLS code in own file
Start to break down the massive ipfixlolib.c into smaller modular
parts so that it's easier to understand.

No functional changes with this commit.

Still quite a lot of conditional compilation for SUPPORT_DLTS in
ipfixlolib.c but it's start.

ipfixlolib_private.h is for declarations that shared between files
within ipfixlolib but not part of the public interface to the rest of
Vermont.
I suspect there is more stuff in ipfixlolib.h that could be moved
private headers.