Simplify the logging code interaction with main
The existing syslog mask and msg level were essentially the same, as
setting the syslog mask was always also setting the msg level. Made
this more obvious by moving all the syslog handling code into the msg
module. This also give clean separation of responsibilities, and
reduces the code size, and makes the signal handlers simpler.

The significant change is dynamically allocate the ConfigManager so
that can be deleted explicitly, instead of when going out scope of
main, thus allowing for msg_shutdown to be called after this, as the
ConfigManager's destructor makes use of messaging. Being able to call
msg_shutdown mean syslog can be flushed/closed and pthread mutex not
left dangling.