Use correct printf format specifiers
use 'clang-tidy' to automatically fix a bunch of incorrect format
specifiers that don't match the arguments passed that compiler was
complaining about. This get us a lot closer to have silent warning
free output during a build.
One day we might even be able to turn on -Werror in CMakeList.txt