Fix building with SUPPORT_DTLS=OFF and libssl-dev not installed
There was still some code that was unconditionally using dtls code,
which requires ssl. This has now been made conditional.

I suspect this is indicative of code smell here, and that perhaps
better encapsulation is needed around this. But that come later as
more attempts are made to modularise the code.

Fixes #85