Don't leak the xml document in the ConfigManager
This actually deletes the object on exit.

Better would be to use proper RAII instead of using pointers for this
class data, but would involve checking if the document is shared, and
if reference or shr_ptr was needed in that case. It's probaby not
shared in which case simply RAII with class data would be work.