Added default values for some variables in the IpfixDbWriter
This commit adds default values for two variables, ipfixId and
enterpriseId, in the IpfixDbWriterSQL. Usually, these variables will
always be set or an exception will be thrown, but gcc (tested with
gcc 8.3.0 on debian) doesn't see this and will fail to compile due
to -Werror. This commit fixes the issue by setting both variables to
zero at initialization.